+++ migrations/0001_create_ab_exposures.sql
@@ -5,6 +5,7 @@ CREATE TABLE IF NOT EXISTS ab_exposures
     subject_id  String,
     is_forced   UInt8 DEFAULT 0,
     is_fallback UInt8 DEFAULT 0,
+    is_sticky   UInt8 DEFAULT 0,
     environment String DEFAULT '',
     ts          DateTime DEFAULT now()
 )